% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvrt_prcnt.R
\name{cnvrt.prcnt}
\alias{cnvrt.prcnt}
\title{Konwersja danych procentowych}
\usage{
cnvrt.prcnt(data)
}
\arguments{
\item{data}{dane w formacie XX,X\% - vector lub matrix}
}
\description{
Konwertuje dane pochodzace z arkuszy Excel z formatu procentowego do formatu numerycznego.
}
