% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_location.R
\name{label_location}
\alias{label_location}
\title{Ustawianie pozycji etykiety względem zakresu danych}
\usage{
label_location(data, ratio = 30)
}
\arguments{
\item{data}{dane do położenia etykiety}

\item{ratio}{proporcja względem maksymalnej wartości danych}
}
\description{
Ustawianie pozycji etykiety względem zakresu danych
}
