% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleta.R
\name{paleta.raporty}
\alias{paleta.raporty}
\title{Paleta kolorow do raportow}
\usage{
paleta.raporty(kolory)
}
\arguments{
\item{kolory}{wektor nazw kolorow do zastosowania (czerwony, zielony, fioletowy, niebieski, granatowy, pomaranczowy, szary, malinowy, zolty)}
}
\description{
Zwraca wektor z wybranymi kolorami z palety stosowanej w raportach.
}
