% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_wrap_sieroty.R
\name{str_wrap_sieroty}
\alias{str_wrap_sieroty}
\title{Zawijanie etykiet na sierotach typograficznych}
\usage{
str_wrap_sieroty(string, width)
}
\arguments{
\item{string}{etykieta}

\item{width}{szerokosc etykiety}
}
\description{
Zawijanie etykiet na sierotach typograficznych
}
