% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tick_marks_mod.R
\name{tick_marks_mod}
\alias{tick_marks_mod}
\title{Dodaje znaczniki na osi pomiędzy etykietami zmiennych}
\usage{
tick_marks_mod(plot, variable, ax = "x", tick.length = 0.01,
  tick.colour = "grey")
}
\arguments{
\item{plot}{wykres z ggplot2}

\item{variable}{zmienna do której mają zostać dodane znaczniki}

\item{ax}{oś na której mają się znaleźć znaczniki}

\item{tick.length}{długość znaczników}

\item{tick.colour}{kolor znacznikóW}
}
\description{
Dodaje znaczniki na osi pomiędzy etykietami zmiennych
}
