% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_plot.R
\name{expand_plot}
\alias{expand_plot}
\title{Rozszerzenie wykresu do rozmiaru obejmującego etykiety}
\usage{
expand_plot(plot, axis = "y")
}
\arguments{
\item{plot}{wykres do rozszerzenia}

\item{axis}{os do rozszerzenia}
}
\description{
Rozszerzenie wykresu do rozmiaru obejmującego etykiety
}
